:recycle: refactor:: Organizacao do codigo e nomeclaturas simplificadas :bulb: docs: modificação no readme.md e implementação de requirements.txt para facilitar instalação das bibliotecas necessarias
