+++ .gitignore
@@ -1,0 +1,7 @@
+node_modules
+__pycache__
+job_vacancies.csv
+embeds.json
+.env
+dist
+package-lock.json
