+++ requirements.txt
@@ -1,0 +1,4 @@
+# Python 3.12.4
+pandas==1.3.3
+openpyxl==3.0.9
+matplotlib==3.4.3
